Detect and work around the following bug in pre-3.0 gccs:
struct X
{
    template <typename T2>
    X operator = (T2 &) {return *this;};
};

template X X::operator=<> (float &);

These compilers reported
  x.cc:7: 'operator =' not defined
  x.cc:7: confused by earlier errors, bailing out
and wanted that we write
  template X X::template operator=<> (float &);
instead. This is not what the standard prescribes, and is also not necessary for regular member functions.

Unfortunately, we cannot simply add that 'template' here, since otherwise Sun's Forte compiler no more
groks this. So we define a symbol DEAL_II_MEMBER_OP_TEMPLATE_INST which is empty for standards
compliant compilers, or 'template' in case of broken gcc's.


git-svn-id: https://svn.dealii.org/trunk@5615 0785d39b-7218-0410-832d-ea1e28bc413d